Real-Time-QA/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi>=0.103.0
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2>=3.1.2
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.15
pydantic>=2.3.0
pydantic-extra-types==2.10.2
pydantic-settings>=2.0.3
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv>=1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn>=0.23.2
watchfiles==1.0.4
websockets>=11.0.3
aiohttp>=3.9.0
